requirements.txt
pyTelegramBotAPI
requests
python-dotenv
